NitinLatest/blogging/requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.6
django-filter==2.2.0
django-rest-swagger==2.1.1
django-unixtimestampfield==0.3.9
djangorestframework==3.11.0
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pylint==2.4.4
pytz==2019.3
PyYAML==5.2
requests==2.22.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.7
wrapt==1.11.2
